LdFlags {RcppThread}	R Documentation

Get portable linker flags for libraries building on RcppThread

Description

To be used in Makevars on Linux and OSX. Returns a string with linker flags for pthread and libatomic, if available.

Usage

LdFlags()

Details

Use as ‘PKG_LIBS = $(R_HOME)/bin/Rscript -e ’RcppThread::LdFlags()'.

[Package RcppThread version 2.3.0 Index]